share funtionality of interfaces
